- moved Common-Interfaces to API - changed VelocityLogger to general Plugin logger
