+++ src/main/java/com/jannik_kuehn/loritime/api/CommonCommand.java
@@ -1,6 +1,5 @@
 package com.jannik_kuehn.loritime.api;
 
-import com.jannik_kuehn.loritime.common.utils.CommonSender;
 import com.velocitypowered.api.command.CommandSource;
 
 import java.util.List;
