+++ src/main/java/com/jannik_kuehn/loritime/api/CommonSender.java
@@ -1,4 +1,4 @@
-package com.jannik_kuehn.loritime.common.utils;
+package com.jannik_kuehn.loritime.api;
 
 import net.kyori.adventure.text.TextComponent;
