+++ src/main/java/com/jannik_kuehn/loritime/common/command/LoriTimeInfoCommand.java
@@ -3,7 +3,7 @@ package com.jannik_kuehn.loritime.common.command;
 import com.jannik_kuehn.loritime.api.CommonCommand;
 import com.jannik_kuehn.loritime.common.LoriTimePlugin;
 import com.jannik_kuehn.loritime.common.config.localization.Localization;
-import com.jannik_kuehn.loritime.common.utils.CommonSender;
+import com.jannik_kuehn.loritime.api.CommonSender;
 import com.velocitypowered.api.command.CommandSource;
 import net.kyori.adventure.text.TextComponent;
 import net.kyori.adventure.text.minimessage.MiniMessage;
